install native: execute Linux logic

+++ dockerfiles/install_dir_domino/install_linux.sh
@@ -44,7 +44,9 @@ install_linux_packages()
     install_package gdb
   else
     install_package gdb-minimal
-    ln -s /usr/bin/gdb.minimal /usr/bin/gdb
+    if [ ! -e /usr/bin/gdb.minimal ]; then
+      ln -s /usr/bin/gdb.minimal /usr/bin/gdb
+    fi
   fi
 
   # SUSE
